Home

SixDOF.jl

SixDOF.aeroforces

SixDOF.inertialtobody

SixDOF.wind

SixDOF.windaxes

SixDOF.aeroforces — Method.

A simple (mostly) linear aerodynamics model

source

SixDOF.inertialtobody — Method.

Construct a rotation matrix from inertial frame to body frame

The assumed order of rotation is

psi radians about the z axis,

theta radians about the y axis,

phi radians about the x axis.

This is an orthogonal transformation so its inverse is its transpose.

source

SixDOF.wind — Method.

Returns

u, v, w: wind velocities in inertial frame

ug, vg, wg: gust velocities in body frame (just a convenience to allow some velocities in body frame)

source

SixDOF.windaxes — Method.

Compute relative velocity in wind axes (airspeed, aoa, sideslip)

source